コミュニティのクライアント系の追加。

+++ src/main/java/io/github/bucheapp/roost/services/ChatServiceImpl.java
@@ -1,5 +1,6 @@
 package io.github.bucheapp.roost.services;
 
+import java.io.IOException;
 import java.time.LocalDateTime;
 
 import jakarta.transaction.Transactional;
@@ -12,9 +13,13 @@ import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.server.ResponseStatusException;
 
 import io.github.bucheapp.roost.dto.request.ChatRequest;
+import io.github.bucheapp.roost.models.ApprovalChat;
 import io.github.bucheapp.roost.models.Chat;
 import io.github.bucheapp.roost.models.ChatType;
 import io.github.bucheapp.roost.models.Community;
+import io.github.bucheapp.roost.models.CommunityState;
+import io.github.bucheapp.roost.models.MediaContent;
+import io.github.bucheapp.roost.models.MediaType;
 import io.github.bucheapp.roost.models.Room;
 import io.github.bucheapp.roost.models.TextChat;
 import io.github.bucheapp.roost.models.User;
@@ -37,6 +42,9 @@ public class ChatServiceImpl implements ChatService {
 	private RoomRepository roomRepository;
 	
 	@Autowired
+	private FileService fileService;
+	
+	@Autowired
 	private AuthContext authContext;
 	
 	@Autowired
@@ -56,7 +64,7 @@ public class ChatServiceImpl implements ChatService {
 
 	@Override
 	@Transactional
-	public Chat createChat(long publicId, ChatRequest req) {
+	public Chat createChat(long publicId, ChatRequest req) throws IOException {
 		long userId = authContext.getCurrentUserId();
 		ChatType type = req.getType();
 		
@@ -67,32 +75,57 @@ public class ChatServiceImpl implements ChatService {
 				.orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, messageUtil.get("user.notfound")));
 		
 		Snowflake snowflake = new Snowflake(workerIdProvider.getWorkerId(), datacenterId);
-		LocalDateTime now = LocalDateTime.now();
 		
 		Chat chat = null;
 		
 		if(type == ChatType.TEXT) {
 			MultipartFile file = req.getFile();
+			String url = null;
 			if(file != null) {
-				checkByte(req.getFile());
-				//TODO: ファイルを作成するロジックを作成
+				fileService.checkByte(file, 10 * 1024 * 1024);
+				
+				if(req.getMediaType() == MediaType.IMAGE) {
+					url = fileService.createImage("image",file);
+				} else if(req.getMediaType() == MediaType.VIDEO) {
+					url = fileService.createVideo("video",file);
+				} else if(req.getMediaType() == MediaType.AUDIO) {
+					url = fileService.createAudio("audio",file);
+				}
 			}
-			chat = new TextChat(req);
+			
+			chat = new TextChat(
+					snowflake.nextId(),
+					LocalDateTime.now(),
+					user,
+					room,
+					req
+					);
+			
+			if(chat instanceof TextChat textChat) {
+				textChat.setMediaContent(
+						new MediaContent(
+								url,
+								req.getMediaType()
+								)
+						);
+			}
+		} else if(type == ChatType.APPROVAL) {
+			chat = new ApprovalChat(
+					snowflake.nextId(),
+					LocalDateTime.now(),
+					user,
+					room
+					);
 		} else {
 			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, messageUtil.get("unknown.chattype"));
 		}
-		
-		chat.setPublicId(snowflake.nextId());
-		chat.setCreatedAt(now);
-		chat.setRoom(room);
-		chat.setCreator(user);
 		
 		return chatRepository.save(chat);
 	}
 
 	@Override
 	@Transactional
-	public Chat updateChat(long publicId, ChatRequest req) {
+	public Chat updateChat(long publicId, ChatRequest req) throws IOException {
 		ChatType type = req.getType();
 		long userId = authContext.getCurrentUserId();
 		
@@ -111,10 +144,27 @@ public class ChatServiceImpl implements ChatService {
 			MultipartFile file = req.getFile();
 			TextChat textChat = (TextChat) chat;
 			textChat.setContent(req.getContent());
+			String url = null;
 			if(file != null) {
-				checkByte(req.getFile());
-				//TODO: ファイルを作成するロジックを作成
+				fileService.checkByte(file, 10 * 1024 * 1024);
+				
+				if(req.getMediaType() == MediaType.IMAGE) {
+					if(textChat.getMediaContent() != null)
+							fileService.deleteFile("image");
+					url = fileService.createImage("image",file);
+				} else if(req.getMediaType() == MediaType.VIDEO) {
+					url = fileService.createVideo("video",file);
+				} else if(req.getMediaType() == MediaType.AUDIO) {
+					url = fileService.createAudio("audio",file);
+				}
 			}
+			
+			textChat.setMediaContent(
+					new MediaContent(
+							url,
+							req.getMediaType()
+							)
+					);
 		} else {
 			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, messageUtil.get("unknown.chattype"));
 		}
@@ -123,7 +173,10 @@ public class ChatServiceImpl implements ChatService {
 		
 		Room room = chat.getRoom();
 		Community community = room.getCommunity();
-		community.checkStateActive();
+		
+		if(community.getState() != CommunityState.ACTIVE) {
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN, messageUtil.get("community.notactive"));
+		}
 		
 		return chatRepository.save(chat);
 	}
@@ -135,6 +188,10 @@ public class ChatServiceImpl implements ChatService {
 		
 		Chat chat = chatRepository.findByPublicId(publicId)
 				.orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, messageUtil.get("chat.notfound")));
+		
+		if(chat instanceof TextChat textChat) {
+			fileService.deleteFile("image/" + textChat.getMediaContent().getMediaContentUrl());
+		}
 		
 		Room room = chat.getRoom();
 		Community community = room.getCommunity();
@@ -148,19 +205,10 @@ public class ChatServiceImpl implements ChatService {
 			throw new ResponseStatusException(HttpStatus.FORBIDDEN, messageUtil.get("cannot.delete.chat"));
 		}
 		
-		community.checkStateActive();
+		if(community.getState() != CommunityState.ACTIVE) {
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN, messageUtil.get("community.notactive"));
+		}
 		
 		chatRepository.delete(chat);
-	}
-
-	@Override
-	public void checkByte(MultipartFile file) {
-		if(file.isEmpty()) {
-			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, messageUtil.get("file.isempty"));
-		}
-		
-		if (file.getSize() > 10 * 1024 * 1024) {
-			throw new ResponseStatusException(HttpStatus.BAD_REQUEST, messageUtil.get("file.size.exceeds"));
-		}
 	}
 }
