+++ src/main/java/io/github/bucheapp/roost/services/RoomServiceImpl.java
@@ -1,20 +1,25 @@
 package io.github.bucheapp.roost.services;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 import jakarta.transaction.Transactional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import io.github.bucheapp.roost.dto.request.RoomRequest;
 import io.github.bucheapp.roost.dto.request.UpdateRoomRequest;
+import io.github.bucheapp.roost.models.Chat;
 import io.github.bucheapp.roost.models.Community;
+import io.github.bucheapp.roost.models.CommunityState;
 import io.github.bucheapp.roost.models.Room;
 import io.github.bucheapp.roost.models.User;
+import io.github.bucheapp.roost.repositories.ChatRepository;
 import io.github.bucheapp.roost.repositories.CommunityRepository;
 import io.github.bucheapp.roost.repositories.RoomRepository;
 import io.github.bucheapp.roost.repositories.UserRepository;
@@ -32,6 +37,9 @@ public class RoomServiceImpl implements RoomService {
 	
 	@Autowired
 	private CommunityRepository communityRepository;
+	
+	@Autowired
+	private ChatRepository chatRepository;
 	
 	@Autowired
 	private WorkerIdProvider workerIdProvider;
@@ -63,15 +71,19 @@ public class RoomServiceImpl implements RoomService {
 		Community community = communityRepository.findByPublicId(publicId)
 				.orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, messageUtil.get("community.notfound")));
 		
-		Snowflake snowflake = new Snowflake(workerIdProvider.getWorkerId(), datacenterId);
-		LocalDateTime now = LocalDateTime.now();
+		if(community.getState() != CommunityState.ACTIVE) {
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN, messageUtil.get("community.notactive"));
+		}
 		
-		Room room = new Room();
-		room.setName(name);
-		room.setCreatedAt(now);
-		room.setPublicId(snowflake.nextId());
-		room.setCreator(user);
-		room.setCommunity(community);
+		Snowflake snowflake = new Snowflake(workerIdProvider.getWorkerId(), datacenterId);
+		
+		Room room = new Room(
+				snowflake.nextId(),
+				name,
+				LocalDateTime.now(),
+				user,
+				community
+				);
 		
 		return roomRepository.save(room);
 	}
@@ -89,7 +101,10 @@ public class RoomServiceImpl implements RoomService {
 		room.setUpdatedAt(now);
 		
 		Community community = room.getCommunity();
-		community.checkStateActive();
+		
+		if(community.getState() != CommunityState.ACTIVE) {
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN, messageUtil.get("community.notactive"));
+		}
 		
 		return roomRepository.save(room);
 	}
@@ -101,8 +116,24 @@ public class RoomServiceImpl implements RoomService {
 				.orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, messageUtil.get("room.notfound")));
 		
 		Community community = room.getCommunity();
-		community.checkStateActive();
+		
+		if(community.getState() != CommunityState.ACTIVE) {
+			throw new ResponseStatusException(HttpStatus.FORBIDDEN, messageUtil.get("community.notactive"));
+		}
 		
 		roomRepository.delete(room);
+	}
+	
+	@Override
+	public List<Chat> getChats(
+			long publicId,
+			Pageable pageable
+			) {
+		Room room = roomRepository.findByPublicId(publicId)
+				.orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND, messageUtil.get("room.notfound")));
+		
+		List<Chat> chats = chatRepository.findByRoom(room, pageable).toList();
+		
+		return chats;
 	}
 }
